refactor(codegen): split emit_module into module.rs (final batch)
Item #2 from doc/plan_compiler_refactor.md, third and final installment.
PR #210 established the pattern with clkgate/regfile/synchronizer; PR #211
moved cam/counter/arbiter/ram/linklist/fsm/pipeline/fifo. This PR finishes
the split by moving `emit_module` and its module-private body-item
helpers to `src/codegen/module.rs`.

Helpers moved (all called only from the emit_module call tree):

- emit_module                 — main module emitter (~550 lines)
- collect_log_files           — log-file path collector for $fwrite
- find_signal_sv_type         — type-of-signal lookup for emit_pipe_reg
- emit_pipe_reg               — pipe_reg port chain expansion
- emit_comb_block             — comb-block always_comb wrapper
- module_has_log              — heuristic for emit_module's $display fd init
- stmt_has_log                — recursive helper for module_has_log
- emit_reg_block              — seq-block always_ff wrapper (~235 lines)
- find_asserted_async_reset   — async-reset detection for emit_reg_block
- emit_latch_block            — latch-block always_latch wrapper
- emit_orphan_reset_regs      — reset-only-assignment emission for regs
                                with reset clauses but no seq-block writes
- emit_reset_only_assignment  — helper for emit_orphan_reset_regs

What stays in `mod.rs`: shared helpers used across construct submodules
— the stmt walker (`emit_stmt`, `emit_if_else`, `emit_for_loop_sv`,
`emit_log_stmt`), pattern + expression emitters, reset / signedness /
indent helpers, and the top-level `emit` dispatch over `Item`.

Final file sizes:
- mod.rs:                  3381 lines (was 9183 pre-split, **−5802 total**)
- module.rs:               1196 lines
- 11 other construct files: 5914 lines combined

Verified end-to-end:
- 221/221 integration tests pass.
- 24 lib unit tests pass.
- Zero snapshot drift across 34 SV-emission snapshots.
- `cargo build` clean.

Item #2 is now complete. The codegen module mirrors `sim_codegen/`'s
per-construct layout: `mod.rs` holds shared helpers and the `Codegen`
struct + dispatch; each construct lives in its own file contributing
`pub(super) fn emit_<name>` to the impl block.
